Added endpoint per issue #71 - adding responsibility to faculty member

+++ WhoteachswhatJPA/src/main/java/com/seneca/repository/ResponsibilityToFacultyDao.java
@@ -8,15 +8,16 @@ public interface ResponsibilityToFacultyDao
 {
    public List<ResponsibilityToFaculty> getAll();
 
-   public ResponsibilityToFaculty getById(Long id);
+	public ResponsibilityToFaculty getById(Integer id);
 
    public void update(ResponsibilityToFaculty entity);
 
    public void create(ResponsibilityToFaculty entity);
 
-   public void delete(Long id);
+	public void delete(Integer id);
 
-   public List<ResponsibilityToFaculty> search(ResponsibilityToFaculty search, Long first, Integer maxItems);
+	public List<ResponsibilityToFaculty> search(ResponsibilityToFaculty search,
+			Integer first, Integer maxItems);
 
    public Long getCount(ResponsibilityToFaculty search);
 }
